vagaPythonSenior/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-redis==5.4.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
drf-yasg==1.21.7
structlog==23.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
dj-database-url==2.1.0 